+++ packages/map-interface/src/location-details/xdd-panel/index.ts
@@ -16,16 +16,30 @@ export interface XDDSnippet {
   highlight: string[];
 }
 
-export function XddExpansion({ xddInfo }) {
+export function XddExpansion({
+  xddInfo,
+  expanded = false,
+  nestedExpanded = true,
+}) {
   return h(xDDPanelCore, {
     className: "regional-panel",
     data: xddInfo,
     isFetching: xddInfo == undefined || xddInfo.length === 0,
+    expanded,
+    nestedExpanded,
   });
 }
 
-export function xDDPanelCore({ isFetching, data: xddInfo, ...rest }) {
+export function xDDPanelCore({
+  isFetching,
+  data: xddInfo,
+  expanded,
+  nestedExpanded,
+  ...rest
+}) {
   const groupedData = groupSnippetsByJournal(xddInfo);
+
+  console.log("expanded", expanded);
 
   return h(
     ExpansionPanel,
@@ -34,12 +48,14 @@ export function xDDPanelCore({ isFetching, data: xddInfo, ...rest }) {
       title: "Primary literature",
       helpText: "via xDD",
       ...rest,
+      expanded,
     },
     [
       h.if(isFetching)(Spinner),
       h.if(!isFetching && xddInfo.length > 0)([
         Array.from(groupedData.entries())?.map(([journal, snippets]) => {
           return h(Journal, {
+            nestedExpanded,
             name: journal,
             articles: snippets,
             publisher: snippets[0].publisher,