+++ packages/feedback-components/CHANGELOG.md
@@ -4,9 +4,24 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [1.1.8] - 2025-07-30
+
+- On load, have any nodes matching autoselect input automically selected
+- View only bugs fixed
+
+## [1.1.7] - 2025-07-25
+
+- Can't select tree in view only mode
+- View matches on hover in view only mode
+
+## [1.1.6] - 2025-07-24
+
+- Fix view and match glitches
+
+## [1.1.5] - 2025-07-24
 
 - Simplified match logic
+- Added view only mode
 
 ## [1.1.4] - 2025-07-16